+++ __tests__/app-5-5-auth.js
@@ -19,51 +19,6 @@ describe('generator-laravel-5:app', () => {
       .on('end', done);
   }, 1200000);
 
-  it('creates files', done => {
-    assert.file(['webpack.mix.js', 'package.json', '.gitignore']);
-    done();
-  });
-
-  it('composer.json contains laravel packages and script', done => {
-    assert.fileContent('composer.json', 'doctrine/dbal');
-    assert.fileContent('composer.json', 'barryvdh/laravel-ide-helper');
-    assert.fileContent('composer.json', 'barryvdh/laravel-debugbar');
-    assert.fileContent('composer.json', 'barryvdh/laravel-cors');
-    assert.fileContent('composer.json', 'phpmetrics/phpmetrics');
-    assert.fileContent('composer.json', 'beyondcode/laravel-self-diagnosis');
-    assert.jsonFileContent('composer.json', {
-      scripts: {
-        analyze: ['phpmetrics --report-html=phpmetrics ./app']
-      }
-    });
-
-    done();
-  });
-
-  it('webpack.mix.js is configured correctly', done => {
-    assert.fileContent(
-      'webpack.mix.js',
-      "mix.js('resources/assets/js/app.js', 'public/js');"
-    );
-    assert.fileContent(
-      'webpack.mix.js',
-      "mix.sass('resources/assets/sass/app.scss', 'public/css');"
-    );
-    assert.fileContent(
-      'webpack.mix.js',
-      'mix.browserSync({\n' +
-        "  proxy: 'http://localhost',\n" +
-        "  host: 'localhost',\n" +
-        '  open: false,\n' +
-        '  watchOptions: {\n' +
-        '    usePolling: true\n' +
-        '  }\n' +
-        '});'
-    );
-
-    done();
-  });
-
   it('creates laravel auth layer files', done => {
     assert.file([
       'app/Http/Controllers/HomeController.php',