+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.5.0
+* Generating ide-helper files.
+* Creating a `database.sqlite` file in the database folder.
+* Integrated [EasyCodingStandard](https://github.com/Symplify/EasyCodingStandard) to check and fix the code style.
+
 ## 1.4.0
 * Optional initialization of a local git repository.
 * Optimized documentation and code comments.